+++ changelog.md
@@ -14,8 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Testing
     * Pages access tests.
     * Requests authentication tests.
-    * User factory.
     * New borrowing validation tests.
+* Models factories
+    * Users.
+    * Inventory items.
 * Register validation
     * Promotion must now be within an acceptable range.
 * Validation